requirements.txt
openai==1.75.0
python-dotenv==1.0.1
opencv-python-headless==4.9.0.80
pillow==10.2.0
tinydb==4.8.0
nicegui==2.17.0
fastapi==0.115.12
uvicorn==0.34.3
slowapi==0.1.9
pydantic==2.11.4
